# swagger/Kyabia.yml
...
swagger: '2.0'
...
info:
  title: Kyabia
  description: |
    Kyabia - **K**araoke **Vi**deo **Ar**biter | API for the Web frontend
  version: "0.1.0"

# Main service information
host: 'localhost:3210'
basePath: '/api'
schemes:
  - 'https'
  - 'http'
consumes:
  - 'application/json'
produces:
  - 'application/json'
...
paths:
...
  /scrapes:
    get:
      tags:
        - 'Admin API'
      description: |
        Retrieves a list with the status of all scraping operations currently 
        queued.
        
        A scraping operation is always identified by the root directory path
        where the operation began. Multiple operations on the same root path
        (or a sub-path of it) are not allowed.
      security:
        - sessionToken: []
      responses:
        200:
          description: 'Successful response'
          schema:
            $ref: '#/definitions/ScrapeListResponse'
        403:
          description: 'Not authorized or insufficient rights'
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  /playlists:
    get:
      tags:
        - 'Admin API'
      description: |
        Gets the list of available playlists.
      security:
        - 'sessionToken': []
      parameters:
        -
          name: includeItems
          in: query
          required: false
          type: boolean
          default: false
      responses:
        200:
          description: 'Successful response'
        403:
          description: 'Not authorized or insufficient rights'
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      tags:
        - 'Admin API'
      description: |
        Creates a new, empty playlist
      security:
        - 'sessionToken': []
      responses:
        200:
          description: 'Contents of the requested playlist'
        403:
          description: 'Not authorized or insufficient rights'
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  DefaultResponse:
    type: object
    required:
      - ok
    properties:
      ok:
        type: boolean
        description: |
          `true` if the call was successful, `false` if not
  ErrorResponse:
    type: object
    allOf:
      - $ref: '#/definitions/DefaultResponse'
    properties:
      error:
        type: string
        description: |
          A machine-readable error code
      errorMessage:
        type: string
        description: |
          A human-readable description of the error
...
  ScrapeListResponse:
    type: object
    allOf:
      - $ref: '#/definitions/DefaultResponse'
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/Scrape'
        description: |
          List of scrapes in the queue
...
  Scrape:
    type: object
    properties:
      status:
        type: string
        description: 'The status of the scrape'
        enum:
          - 'queued'
          - 'running'
          - 'finished'
          - 'failed'
          - 'cancelled'
      rootDir:
        type: string
        description: 'The root directory the scrape started at'
      currentDir:
        type: string
        description: 'The current directory the scrape is working on'
      currentFile:
        type: string
        description: 'The current file the scrape is working on'
      filesScraped:
        type: number
        format: int64
        description: 'The number of files already scraped'
      newFiles:
        type: number
        format: int64
        description: |
          The number of files scraped that were added to the database
      updatedFiles:
        type: number
        format: int64
        description: |
          The number of files scraped that were updated in the database
      startedAt:
        type: string
        format: date-time
        description: |
          The timestamp when the scrape was started
      error:
        type: string
        description: |
          If the scrape failed, this is the error that caused it
    

# Available security types                  
securityDefinitions:
  sessionToken:
    type: apiKey
    in: header
    name: token